+++ Project/src/MakeCall/CallCard.js
@@ -12,6 +12,7 @@ import LocalVideoPreviewCard from './LocalVideoPreviewCard';
 import { Dropdown } from 'office-ui-fabric-react/lib/Dropdown';
 import { LocalVideoStream, Features, LocalAudioStream } from '@azure/communication-calling';
 import { utils } from '../Utils/Utils';
+import VolumeVisualizer from "./VolumeVisualizer";
 
 export default class CallCard extends React.Component {
     constructor(props) {
@@ -21,12 +22,6 @@ export default class CallCard extends React.Component {
         this.deviceManager = props.deviceManager;
         this.state = {
             callState: this.call.state,
-            localVolumeLevel: 0,
-            remoteVolumeLevel: 0,
-            localVolumeLevelSubscription: undefined,
-            remoteVolumeLevelSubscription: undefined,
-            localVolumeIndicator: undefined,
-            remoteVolumeIndicator: undefined,
             callId: this.call.id,
             // localParticipantid: this.call.feature(Features.DebugInfo).localParticipantId,
             remoteParticipants: this.call.remoteParticipants,
@@ -119,36 +114,12 @@ export default class CallCard extends React.Component {
                 });
             });
 
-            let localVolumeStateSetter = undefined;
-            let handleSelectedMicrophoneVolumeSubscription = async () => {        
-                let localVolumeIndicator = await (new LocalAudioStream(this.deviceManager.selectedMicrophone).getVolume());
-                localVolumeStateSetter = ()=>{
-                    this.setState({ localVolumeLevel: localVolumeIndicator.level });
-                }
-                localVolumeIndicator.on('levelChanged', localVolumeStateSetter);
-                this.setState({ localVolumeLevelSubscription: localVolumeStateSetter });
-                this.setState({ localVolumeIndicator: localVolumeIndicator });                             
-            }
-            handleSelectedMicrophoneVolumeSubscription();
-
-            let remoteVolumeStateSetter = undefined;
-            let handleRemoteVolumeSubscription = async () => {                
-                let remoteVolumeIndicator = await this.call.remoteAudioStreams[0].getVolume();
-                remoteVolumeStateSetter = ()=>{
-                    this.setState({ remoteVolumeLevel: remoteVolumeIndicator.level });
-                }
-                remoteVolumeIndicator.on('levelChanged', remoteVolumeStateSetter);
-                this.setState({ remoteVolumeLevelSubscription: remoteVolumeStateSetter });
-                this.setState({ remoteVolumeIndicator: remoteVolumeIndicator });                              
-            }
-
             this.deviceManager.on('selectedSpeakerChanged', () => {
                 this.setState({ selectedSpeakerDeviceId: this.deviceManager.selectedSpeaker?.id });
             });
 
             this.deviceManager.on('selectedMicrophoneChanged', () => {
                 this.setState({ selectedMicrophoneDeviceId: this.deviceManager.selectedMicrophone?.id });
-                handleSelectedMicrophoneVolumeSubscription();
             });
 
             const callStateChanged = () => {
@@ -162,9 +133,6 @@ export default class CallCard extends React.Component {
                         this.callFinishConnectingResolve();
                     }
                 }
-                if (this.call.state === 'Connected'){
-                    this.call.on('remoteAudioStreamsUpdated', handleRemoteVolumeSubscription)
-                }
                 if (this.call.state === 'Incoming') {
                     this.setState({ selectedCameraDeviceId: cameraDevices[0]?.id });
                     this.setState({ selectedSpeakerDeviceId: speakerDevices[0]?.id });
@@ -173,7 +141,6 @@ export default class CallCard extends React.Component {
 
                 if (this.call.state === 'Disconnected') {
                     this.setState({ dominantRemoteParticipant: undefined });
-                    this.componentWillUnmount()
                 }
             }
             callStateChanged();
@@ -317,11 +284,6 @@ export default class CallCard extends React.Component {
             this.call.feature(Features.LiveStream).on('liveStreamsUpdated', (streams) => streamUpdatedHandler(streams, 'Live'));
             this.call.feature(Features.ComposedStream).on('composedStreamsUpdated', (streams) => streamUpdatedHandler(streams, 'Composed'));
         }
-    }
-
-    async componentWillUnmount() {
-        this.state.localVolumeIndicator.off('levelChanged', this.state.localVolumeLevelSubscription);
-        this.state.remoteVolumeIndicator.off('levelChanged', this.state.remoteVolumeLevelSubscription);
     }
 
     subscribeToRemoteParticipant(participant) {
@@ -659,22 +621,8 @@ export default class CallCard extends React.Component {
                     }
                     <div className={this.state.callState === 'Connected' ? `ms-Grid-col ms-sm12 ms-lg12 ms-xl12 ms-xxl9` : 'ms-Grid-col ms-sm12 ms-lg12 ms-xl12 ms-xxl12'}>
                     {
-                        (this.state.callState === 'Connected') && <div className="volume-indicatordiv">
-                                        {
-                                            (this.state.callState === 'Connected' && !this.state.micMuted) &&
-                                            <div className='elements'>
-                                            <label>Remote Volume Visualizer</label>
-                                            <div className="volumeVisualizer" style={{'--volume':this.state.remoteVolumeLevel + '%'}}></div>
-                                        </div>
-                                        }
-                                        {
-                                            (this.state.callState === 'Connected' && !this.state.micMuted) &&
-                                            <div className='elements'>
-                                            <label>Selected Microphone Volume Visualizer</label>
-                                            <div className="volumeVisualizer" style={{'--volume':this.state.localVolumeLevel + '%'}}></div>
-                                        </div>
-                                        }
-                        </div>
+                        (this.state.callState === 'Connected') && !this.state.micMuted &&
+                        <VolumeVisualizer call={this.call} deviceManager={this.deviceManager} />
                     }
                         <div className="mb-2">
                             {
